+++ Sources/KeyboardKit/Styles/InputCalloutStyle.swift
@@ -9,16 +9,19 @@
 import SwiftUI
 
 /**
- This can be used to style `InputCallout` views, by applying
- the `inputCalloutStyle` view modifier.
+ This view style can be used to style ``InputCallout`` views.
  
- You can modify the `.standard` style instance to change the
- standard, global style.
+ You can modify ``InputCallout/standard`` to edit the global
+ style of an input callout.
  
- Note that `calloutSize` only applies to portrait when on an
- iPhone. In landscape, `InputCallout` will enforce a smaller
- callout size, since a callout can't expand beyond the edges
- of a keyboard extension.
+ The ``calloutSize`` specifies a **minimum** size to use for
+ the callout. If other factors, like button size, curve size,
+ padding etc. requires the callout to be larger, the size is
+ ignored and the required size will be applied.
+ 
+ Note that the ``calloutSize`` height will be ignored when a
+ phone displays a callout in landscape, since callouts can't
+ expand beyond the edges of a keyboard extension.
  */
 public struct InputCalloutStyle {
     
@@ -27,14 +30,17 @@ public struct InputCalloutStyle {
      
      - Parameters:
        - callout: The callout style to use, by default `.standard`.
-       - calloutSize: The size of the callout bubble.
+       - calloutPadding: The padding to apply to the callout content.
+       - calloutSize: The minimum size of the callout bubble.
        - font: The font to use in the callout.
      */
     public init(
         callout: CalloutStyle = .standard,
-        calloutSize: CGSize = CGSize(width: 65, height: 55),
+        calloutPadding: EdgeInsets = EdgeInsets(top: 2, leading: 2, bottom: 2, trailing: 2),
+        calloutSize: CGSize = CGSize(width: 0, height: 55),
         font: Font = Font.largeTitle.weight(.light)) {
         self.callout = callout
+        self.calloutPadding = calloutPadding
         self.calloutSize = calloutSize
         self.font = font
     }
@@ -45,10 +51,15 @@ public struct InputCalloutStyle {
     public var callout: CalloutStyle
     
     /**
-     The size of the callout above the button area.
+     The padding to apply to the callout content.
+     */
+    public var calloutPadding: EdgeInsets
+    
+    /**
+     The minimum size of the callout above the button area.
      
-     If the width is smaller than the button width, then the
-     `InputCallout` applies `buttonSize` plus padding.
+     If other factors, like button size, curve size, padding
+     etc. requires the callout to be larger, this is ignored.
      */
     public var calloutSize: CGSize
     
@@ -73,7 +84,7 @@ extension InputCalloutStyle {
      */
     static var preview1 = InputCalloutStyle(
         callout: .preview1,
-        calloutSize: CGSize(width: 60, height: 40),
+        calloutSize: CGSize(width: 0, height: 40),
         font: .body)
     
     /**
@@ -81,6 +92,7 @@ extension InputCalloutStyle {
      */
     static var preview2 = InputCalloutStyle(
         callout: .preview2,
-        calloutSize: CGSize(width: 10, height: 60),
+        calloutPadding: EdgeInsets(top: 10, leading: 20, bottom: 30, trailing: 40),
+        calloutSize: CGSize(width: 100, height: 60),
         font: .footnote)
 }