+++ RELEASE_NOTES.md
@@ -10,6 +10,12 @@ Breaking changes can still occur in minor versions and patches, though, if the a
 
 
 
+## 5.2
+
+This version bumps the package Swift version to 5.5 to enable extended DocC support.
+
+
+
 ## 5.1
 
 This version adds support for `estonian`, `latvian` and `lithuanian`.